+++ pkg/extension/sumologicextension/extension_test.go
@@ -1377,6 +1377,53 @@ func TestRegistrationRequestPayload(t *testing.T) {
 	require.NoError(t, se.Shutdown(context.Background()))
 }
 
+func TestWatchCredentialKey(t *testing.T) {
+	cfg := createDefaultConfig().(*Config)
+	cfg.Credentials.InstallToken = "dummy_install_token"
+	se, err := newSumologicExtension(cfg, zap.NewNop(), component.NewID("sumologic"), "1.0.0")
+	require.NoError(t, err)
+
+	ctx := context.Background()
+	ctxc, cancel := context.WithCancel(ctx)
+	cancel()
+	v := se.WatchCredentialKey(ctxc, "")
+	require.Equal(t, v, "")
+
+	v = se.WatchCredentialKey(context.Background(), "foobar")
+	require.Equal(t, v, "")
+
+	go func() {
+		time.Sleep(time.Millisecond * 100)
+		se.registrationInfo.CollectorCredentialKey = "test-credential-key"
+		close(se.credsNotifyUpdate)
+	}()
+
+	v = se.WatchCredentialKey(context.Background(), "")
+	require.Equal(t, v, "test-credential-key")
+}
+
+func TestCreateCredentialsHeader(t *testing.T) {
+	cfg := createDefaultConfig().(*Config)
+	cfg.Credentials.InstallToken = "dummy_install_token"
+	se, err := newSumologicExtension(cfg, zap.NewNop(), component.NewID("sumologic"), "1.0.0")
+	require.NoError(t, err)
+
+	_, err = se.CreateCredentialsHeader()
+	require.Error(t, err)
+
+	se.registrationInfo.CollectorCredentialId = "test-credential-id"
+	se.registrationInfo.CollectorCredentialKey = "test-credential-key"
+
+	h, err := se.CreateCredentialsHeader()
+	require.NoError(t, err)
+
+	authHeader := h.Get("Authorization")
+	token := base64.StdEncoding.EncodeToString(
+		[]byte("test-credential-id:test-credential-key"),
+	)
+	assert.Equal(t, "Basic "+token, authHeader)
+}
+
 func TestGetHostIpAddress(t *testing.T) {
 	ip, err := getHostIpAddress()
 	require.NoError(t, err)